pyproject.toml
[tool.poetry]
name = "management-system"
version = "0.1.0"
description = ""
authors = ["JHoo0118 <kik3078@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pillow = "^10.0.1"
pdf2docx = "^0.5.6"
pandas = "^2.1.1"
openpyxl = "^3.1.2"
docx = "^0.2.4"
simple-term-menu = "^1.6.1"
pyfiglet = "^1.0.2"
termcolor = "^2.3.0"
pypdf = "^3.16.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
